+++ packages/cli/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - Updated project `language` to be automatically inferred when the value is unknown, based on the
   existence of config files (`package.json` = javascript, `tsconfig.json` = typescript).
+- Updated the `InstallNodeDeps` action to install dependencies when a `package.json` change is
+  detected.
 - Added a `moon dep-graph` command for displaying the entire dependency graph in DOT format.
 - Added `--language` and `--type` filter options to `moon query projects`.
 - Added `$language`, `$projectType`, and `$taskType` token variables.